+++ Assignment/Node.cs
